move action enum file to constants dir
